Modify top level directory structure for a more OTP friendly design. Update sync utility to search around for Emakefiles, simplifying compilation.

+++ Quickstart/quickstart_mochiweb.sh
@@ -1,30 +1,25 @@
-#!/bin/sh
-export NITROGEN_HOME=..
-export MOCHIWEB_HOME=../http_servers/mochiweb
-export SIMPLEBRIDGE_HOME=../deps/simple_bridge
+#!/usr/bin/env sh
 cd `dirname $0`
 
-echo Creating link to Nitrogen support files...
+# Check if Mochiweb exists..
+if [ ! -d "../apps/mochiweb" ]; then
+    echo "Could not find Mochiweb in '../apps/mochiweb'."
+    echo "Run ../apps/get_mochiweb.sh to download from http://mochiweb.googlecode.com"
+    exit
+fi
+
+# Compile all required projects...
+(cd ..; make compile)
+(cd ../apps/mochiweb; make all)
+
+# Link to support files...
+echo Creating link to nitrogen support files...
 rm -rf wwwroot/nitrogen
-ln -s ../$NITROGEN_HOME/www wwwroot/nitrogen
+ln -s ../../apps/nitrogen/www wwwroot/nitrogen
 
-echo Compile Nitrogen...
-make -C $NITROGEN_HOME
-
-echo Compile Mochiweb...
-make -C $MOCHIWEB_HOME
-
-echo Compile Simple Bridge...
-make -C $SIMPLEBRIDGE_HOME
-
+# Start Nitrogen on Mochiweb...
 echo Starting Nitrogen on Mochiweb...
-exec erl \
-	-name nitrogen_mochiweb@127.0.0.1 \
-	-pa $PWD/apps $PWD/ebin $PWD/include \
-	-pa $NITROGEN_HOME/ebin $NITROGEN_HOME/include \
-	-pa $SIMPLEBRIDGE_HOME/ebin $SIMPLEBRIDGE_HOME/include \
-	-pa $MOCHIWEB_HOME/ebin $MOCHIWEB_HOME/include \
-	-s make all \
-	-eval "application:start(quickstart_mochiweb)"
-
-
+erl \
+    -name nitrogen_mochiweb@127.0.0.1 \
+    -pa ./ebin ../apps/*/ebin ../apps/*/include \
+    -eval "application:start(quickstart_mochiweb)"
